medicheck-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
psycopg[binary]==3.2.3
PyJWT==2.10.0
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
